Add note about anchor tags not being proxied
Hi there,

Hope this is helpful for other people. It's not immediately clear from the docs that when clicking on a relative link, to initiate an Oauth2 flow for example, the proxy doesn't work because of the Accept header is `text/html`.

Happy to re-word this to make it more clear. Let me know and I can fix.

Thanks!